Revert "refactor: extract service modules"
