+++ src/components/trips/FuelRateSelector.tsx
@@ -50,7 +50,7 @@ const FuelRateSelector: React.FC<FuelRateSelectorProps> = ({
   useEffect(() => {
     const fetchTrips = async () => {
       try {
-        const { getTrips } = await import('../../services/trips');
+        const { getTrips } = await import('../../utils/storage');
         const tripsData = await getTrips();
         setTrips(Array.isArray(tripsData) ? tripsData : []);
       } catch (error) {
