+++ src/components/vehicles/DocumentSummaryPanel.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect, useRef, useMemo } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { X, User, Truck, Calendar, FileText, Shield, Download, Printer as Print, Search, ChevronDown, ChevronUp, Clock, Info, BarChart2, Database, IndianRupee, Bell, FileCheck, AlertCircle, ArrowLeft, ArrowRight } from 'lucide-react';
 import { Vehicle } from '../../types';
-import { getVehicles } from '../../services/vehicles';
+import { getVehicles } from '../../utils/storage';
 import Button from '../ui/Button';
 import Input from '../ui/Input';
 import Select from '../ui/Select';
