+++ src/components/vehicles/VehicleForm.tsx
@@ -3,7 +3,7 @@ import { useForm, Controller, useFieldArray } from 'react-hook-form';
 import { Vehicle, Driver } from '../../types';
 import { getMaterialTypes, MaterialType } from '../../utils/materialTypes';
 import { getReminderContacts, ReminderContact } from '../../utils/reminderService';
-import { getDrivers } from '../../services/drivers';
+import { getDrivers } from '../../utils/storage';
 import { supabase } from '../../utils/supabaseClient';
 import Input from '../ui/Input';
 import Select from '../ui/Select';
