+++ src/pages/AIAlertsPage.tsx
@@ -2,9 +2,7 @@ import React, { useState, useEffect, useMemo } from 'react';
 import Layout from '../components/layout/Layout';
 import { AIAlert } from '../types'; 
 import { getAIAlerts, processAlertAction, runAlertScan } from '../utils/aiAnalytics';
-import { getVehicle, getVehicles } from '../services/vehicles';
-import { getDrivers } from '../services/drivers';
-import { getTrips } from '../services/trips';
+import { getVehicle, getVehicles } from '../utils/storage';
 import DriverAIInsights from '../components/ai/DriverAIInsights';
 import { AlertTriangle, CheckCircle, XCircle, Bell, Search, ChevronRight, BarChart2, Filter, RefreshCw, Truck, Calendar, Fuel, TrendingDown, FileX, PenTool as Tool } from 'lucide-react';
 import Input from '../components/ui/Input';
