+++ src/pages/DriversPage.tsx
@@ -1,8 +1,13 @@
 import React, { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import Layout from "../components/layout/Layout";
-import { getTrips } from "../services/trips";
-import { getDrivers, createDriver, updateDriver, uploadDriverPhoto } from "../services/drivers";
+import { 
+  getDrivers, 
+  getTrips, 
+  createDriver, 
+  updateDriver, 
+  uploadDriverPhoto 
+} from "../utils/storage"; // ⚠️ Confirm field refactor here
 import { supabase } from "../utils/supabaseClient";
 import {
   User, Users,
