+++ src/pages/TripDetailsPage.tsx
@@ -5,10 +5,7 @@ import Layout from '../components/layout/Layout';
 import TripDetails from '../components/trips/TripDetails';
 import TripForm from '../components/trips/TripForm';
 import { Trip, TripFormData, Vehicle, Driver, Destination, Warehouse } from '../types';
-import { getDestination, getWarehouse, getDestinations, getWarehouses } from '../utils/storage';
-import { getVehicle, getVehicles } from '../services/vehicles';
-import { getDriver, getDrivers } from '../services/drivers';
-import { getTrip, updateTrip, deleteTrip, getTrips } from '../services/trips';
+import { getTrip, getVehicle, getDriver, getDestination, updateTrip, deleteTrip, getWarehouse, getTrips, getVehicles, getDrivers, getDestinations, getWarehouses } from '../utils/storage';
 import { getDestinationByAnyId } from '../utils/storage';
 import { getMaterialTypes, MaterialType } from '../utils/materialTypes';
 import { getAIAlerts, AIAlert } from '../utils/aiAnalytics';
