+++ src/pages/TripPnlReportsPage.tsx
@@ -2,10 +2,7 @@ import React, { useState, useEffect, useMemo } from 'react';
 import { useNavigate } from 'react-router-dom';
 import Layout from '../components/layout/Layout';
 import { Trip, Vehicle, Driver, Warehouse } from '../types';
-import { getWarehouses } from '../utils/storage';
-import { getTrips } from '../services/trips';
-import { getVehicles } from '../services/vehicles';
-import { getDrivers } from '../services/drivers';
+import { getTrips, getVehicles, getDrivers, getWarehouses } from '../utils/storage';
 import { 
   Calendar, 
   ChevronDown,
