+++ src/pages/TripsPage.tsx
@@ -11,10 +11,7 @@ import Select from '../components/ui/Select';
 import MultiSelect from '../components/ui/MultiSelect';
 import Checkbox from '../components/ui/Checkbox';
 import { Trip, TripFormData, Vehicle, Driver, Warehouse } from '../types';
-import { getWarehouses } from '../utils/storage';
-import { getTrips, createTrip } from '../services/trips';
-import { getVehicles } from '../services/vehicles';
-import { getDrivers } from '../services/drivers';
+import { getTrips, getVehicles, getDrivers, createTrip, getWarehouses } from '../utils/storage';
 import { getMaterialTypes, MaterialType } from '../utils/materialTypes';
 import { validateTripSerialUniqueness } from '../utils/tripSerialGenerator';
 import { uploadFilesAndGetPublicUrls } from '../utils/supabaseStorage';
