+++ src/pages/admin/AdminDashboard.tsx
@@ -2,8 +2,7 @@ import React, { useState, useEffect } from 'react';
 import Layout from '../../components/layout/Layout';
 import { Settings, Users, Truck, PenTool as Tool, MapPin, AlertTriangle, Bell, FileText, Calendar, FileCheck, BarChart2, Database, MessageSquare, Activity, ShieldCheck } from 'lucide-react'; 
 import { Link } from 'react-router-dom';
-import { getVehicles } from '../../services/vehicles';
-import { getDrivers } from '../../services/drivers';
+import { getVehicles, getDrivers } from '../../utils/storage';
 import { Vehicle, Driver } from '../../types';
 import DataTable from '../../components/shared/DataTable';
 import Button from '../../components/ui/Button';
