+++ src/pages/admin/AdminTripsPage.tsx
@@ -5,10 +5,7 @@ import TripsTable from '../../components/admin/TripsTable';
 import TripsSummary from '../../components/admin/TripsSummary';
 import ExportOptionsModal, { ExportOptions } from '../../components/admin/ExportOptionsModal';
 import { Trip, Vehicle, Driver, Warehouse } from '../../types';
-import { getWarehouses } from '../../utils/storage';
-import { getTrips, updateTrip } from '../../services/trips';
-import { getVehicles } from '../../services/vehicles';
-import { getDrivers } from '../../services/drivers';
+import { getTrips, getVehicles, getDrivers, getWarehouses, updateTrip } from '../../utils/storage';
 import { generateCSV, downloadCSV, parseCSV } from '../../utils/csvParser';
 import { supabase } from '../../utils/supabaseClient';
 import { format, subDays, startOfWeek, endOfWeek, startOfMonth, endOfMonth,
