+++ src/pages/drivers/DriverInsightsPage.tsx
@@ -47,9 +47,7 @@ import Input from "../../components/ui/Input";
 import Select from "../../components/ui/Select";
 import StatCard from "../../components/ui/StatCard";
 import { cn } from "../../utils/cn";
-import { getVehicles } from "../../services/vehicles";
-import { getDrivers } from "../../services/drivers";
-import { getTrips } from "../../services/trips";
+import { getDrivers, getTrips, getVehicles } from "../../utils/storage";
 import type { Driver, Trip, Vehicle } from "../../types";
 
 // Interface for driver performance metrics
